+++ test/expected/foo.js
@@ -1,4 +1,0 @@
-var foo = {
-	bar: 1,
-	baz: 2
-};